DeserializerTests: Restructure test folder layout
